requirements.txt
streamlit
tensorflow
opencv-python-headless
Pillow
numpy
plotly
st-supabase-connection
bcrypt
streamlit-cookies-manager
gdown